requirements.txt
flask==2.3.2
python-dotenv==1.0.0
openai==1.5.0
flask-cors==3.0.10
gunicorn==21.2.0